Fixed method getBoolean()
